Icon loaded on App component first render and implemented basic drag and drop prototype
Images are now loaded on the first render of App component. This list of images are passed to
the Sidebar, IconList, and Arena components. The list of images should also be passed to the
IconPreview component that hasn't been made yet. The drag and drop implemetation is just a test
prototype. Icons that are dragged into the Arena should spawn a draggable element based on where it
was dropped which the drop event can determine. All draggable element states should be saved and
be integrated with socket.io so it syncs for all clients.