Added multistep feature that allows multiple arenas editing
Added Multistep component that displays the number of arenas as steps. On
clicking one of these steps allows the user to switch to that arena state.
Multistep component also has the + and delete (trash icon) button for
adding and deleting an arena state. All icon spawning, moving, editing are
synced between all different arenas/steps. Adding and deleting an arena
are also synced. Further testing is needed to ensure it's functioning
correctly as well as adding and removing log, warning, and error logs.