New users now get the arenaStates from the oldest user
Users who join an existing room will now sync their
arenaStates to the oldest user in the room. This is done
via a few messages between the new user client, server, and
oldest user client. If a room is empty, no arenaStates
are synced, and the first user to join will have an
fresh arena.